+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 #### New
 
 * The container can now resolve parameters with the help of custom injector attributes. The following attributes are included by default:
-	- `InjectConfig` which allows you to inject a config value
-	- `InjectEnv` which allows you to inject a environment variable value
+	- `InjectConfig` which allows you to inject a config value.
+	- `InjectEnv` which allows you to inject a environment variable value.
 
 --------------------------------------------------------