+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 #### New
 
 * The container can now resolve parameters with the help of custom injector attributes. The following attributes are included by default:
+	- `InjectConnection` which allows you to inject a database connection.
 	- `InjectConfig` which allows you to inject a config value.
 	- `InjectEnv` which allows you to inject a environment variable value.
 * Added `disableAfterSunset` option to the `Deprecated` middleware.