+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 #### New
 
 * The container can now resolve parameters with the help of custom injector attributes. The following attributes are included by default:
-	- `InjectConnection` which allows you to inject a database connection.
-	- `InjectConnection` which allows you to inject a Redis connection.
+	- `InjectConnection` which allows you to inject a specific database connection.
+	- `InjectConnection` which allows you to inject a specific Redis connection.
 	- `InjectConfig` which allows you to inject a config value.
 	- `InjectEnv` which allows you to inject a environment variable value.
 * Added `disableAfterSunset` option to the `Deprecated` middleware.