+++ CHANGELOG.md
@@ -199,7 +199,7 @@ The major version bump is due to upping the required PHP version from `8.1` to `
 * Renamed the `ErrorHandler::handle()` method to `ErrorHandler::addHandler()`.
 * Renamed the `ErrorHandler::handler()` method to `ErrorHandler::handle()`.
 * Renamed the `mako\cli\output\helpers` namespace to `mako\cli\output\components`.
-* The Redis cache `clear` implementation will no longer flush the entire database but instead just deleted the cached keys.
+* The Redis cache `clear` implementation will no longer flush the entire database but instead just delete the cached keys.
 
 #### Deprecations